+++ src/Tests/VirtualMachine.fs
@@ -56,4 +56,37 @@ let tests = testList "Virtual Machine" [
         arm { add_resource (vm { name "foo"; username "foo"; custom_script "foo"; custom_script_files [ "http://test.fsx" ] }) } |> ignore
         arm { add_resource (vm { name "foo"; username "foo"; custom_script "foo" }) } |> ignore
     }
+
+    test "CustomData is correctly encoded" {
+        let deployment =
+            arm {
+                add_resources
+                    [ vm { name "foo"; username "foo"; custom_data "foo"} ]
+            }
+        let json = deployment.Template |> Writer.toJson
+        let jobj = Newtonsoft.Json.Linq.JObject.Parse(json)
+        let customData = jobj.SelectToken("resources[?(@.name=='foo')].properties.osProfile.customData")
+        let actualCustomData = (customData.ToString())
+        let expectedCustomData = "Zm9v"
+        Expect.equal actualCustomData expectedCustomData "customData was not correctly encoded"
+    }
+
+    test "Can remove public Ip" {
+        let deployment =
+            arm {
+                add_resources
+                    [ vm { name "foo"; username "foo"; public_ip None} ]
+            }
+        let json = deployment.Template |> Writer.toJson
+        let jobj = Newtonsoft.Json.Linq.JObject.Parse(json)
+        
+        let publicIps= jobj.SelectTokens("resources[?(@.type=='Microsoft.Network/publicIPAddresses')]")
+        Expect.isEmpty publicIps "No public IP should be created"
+        
+        let nicDependsOn = jobj.SelectTokens("resources[?(@.type=='Microsoft.Network/networkInterfaces')].dependsOn.[*]") |> Seq.map (fun x -> x.ToString())
+        Expect.isEmpty (nicDependsOn |> Seq.filter (fun x->x.Contains("Microsoft.Network/publicIPAddresses"))) "Network Interface should not depende on any public IP"
+        
+        let nicPublicIp = jobj.SelectTokens("resources[?(@.type=='Microsoft.Network/networkInterfaces')].properties.publicIpAddress")
+        Expect.isEmpty (nicPublicIp) "Network Interface should not link to any public IP"
+    }
 ]